Build fixes for Linux and OS X.

+++ boot/boot.cpp
@@ -90,7 +90,7 @@ namespace KrollBoot
 		if (app->HasArgument("debug"))
 		{
 			vector<SharedComponent> resolved = app->GetResolvedComponents();
-			for (int i = 0; i < resolved.size(); i++)
+			for (size_t i = 0; i < resolved.size(); i++)
 			{
 				SharedComponent c = resolved[i];
 				std::cout << "Resolved: (" << c->name << " "
